backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
opencv-python-headless==4.8.1.78
imagehash==4.3.1
faiss-cpu==1.7.4
Pillow==10.1.0
numpy==1.26.2
python-multipart==0.0.6
aiofiles==23.2.1
aiosqlite==0.19.0
httpx==0.25.2
python-dotenv==1.0.0
